requirements.txt
adbutils==1.2.9
apkutils2==1.0.0
Appium-Python-Client==2.10.1
APScheduler==3.10.4
args==0.1.0
asgiref==3.7.2
asn1crypto==1.5.1
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cached-property==1.5.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cigam==0.0.3
click==8.1.3
clint==0.5.1
colorama==0.4.6
decorator==5.1.1
Deprecated==1.2.13
deprecation==2.1.0
Django==4.2.4
django-apscheduler==0.6.2
django-crontab==0.7.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.12.0
Flask==2.3.2
google-play-scraper==1.2.4
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
logzero==1.7.0
lxml==4.9.3
MarkupSafe==2.1.2
mysqlclient==2.2.0
numpy==1.26.1
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.1.2
Pillow==9.5.0
progress==1.6
py==1.11.0
pyaxmlparser==0.3.28
pycparser==2.21
pyelftools==0.29
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
retry==0.9.2
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
tzlocal==5.2
uiautomator==1.0.2
uiautomator2==2.16.23
UIAutomatorViewer==0.0.4
urllib3==2.0.4
webdriver-manager==4.0.0
Werkzeug==2.3.4
whichcraft==0.6.1
wrapt==1.15.0
wsproto==1.2.0
xlrd3==1.1.0
xmltodict==0.13.0
zipp==3.15.0
